Updated README.md to reflect project scope
Updated the README to focus clearly on the scope of flashstache for simple automation of data collection and consolidation of default dashboards.  This is to make sure it's clear that FlashStache is not designed to be a front-end for grafana, but rather simplify data collection for use with grafana capabilities.