Bug Fix:  setup.sh; Issue 6 Enhancement: setup.sh; Issue 4
Issue 4:
Changed reference to 'pip' to 'pip2' to ensure the proper version of Python is used.

Issue 6:
Invalid variable was referenced "pwd",
this was causing configuration changes to .ini files to use
an invalid path of '/flasharray' and other invalid paths.

The end result is that Nginx uwsgi was not working.